Refactor the training score data
I refactored the training score data to make it easier
to expand on the funcionality.
