+++ tracking-analyzer/filesystem.h
@@ -1,0 +1,11 @@
+#ifndef ANALYZER_FILESYSTEM_H
+#define ANALYZER_FILESYSTEM_H
+
+#include <QString>
+
+namespace analyzer
+{
+  auto make_absolute_path(const QString& path) -> QString;
+}  // namespace analyzer
+
+#endif
