+++ tracking-analyzer/gui/main_window.cpp
@@ -1,4 +1,5 @@
 #include "main_window.h"
+#include "filesystem.h"
 #include "tracking_results.h"
 #include "ui_main_window.h"
 #include <QFileDialog>
